gatekeep pool price fetch on contextMatchesParams
