+++ src/App/hooks/usePoolMetadata.ts
@@ -79,6 +79,7 @@ export function usePoolMetadata() {
         defaultRangeWidthForActivePool,
         currentPoolPriceTick,
         activeTab,
+        contextMatchesParams,
     } = useContext(TradeDataContext);
 
     const {
@@ -107,31 +108,6 @@ export function usePoolMetadata() {
 
     const ticksInParams =
         pathname.includes('lowTick') && pathname.includes('highTick');
-
-    // hook to sync token addresses in RTK to token addresses in RTK
-    const contextMatchesParams = useMemo(() => {
-        let matching = false;
-        const tokenAAddress = tokenA.address;
-        const tokenBAddress = tokenB.address;
-
-        if (pathname.includes('tokenA') && pathname.includes('tokenB')) {
-            const getAddrFromParams = (token: string) => {
-                const idx = pathname.indexOf(token);
-                const address = pathname.substring(idx + 7, idx + 49);
-                return address;
-            };
-            const addrTokenA = getAddrFromParams('tokenA');
-            const addrTokenB = getAddrFromParams('tokenB');
-            if (
-                addrTokenA.toLowerCase() === tokenAAddress.toLowerCase() &&
-                addrTokenB.toLowerCase() === tokenBAddress.toLowerCase()
-            ) {
-                matching = true;
-            }
-        }
-
-        return matching;
-    }, [pathname, tokenA, tokenB]);
 
     const baseTokenAddress = useMemo(
         () => sortBaseQuoteTokens(tokenA.address, tokenB.address)[0],
@@ -844,7 +820,6 @@ export function usePoolMetadata() {
     }, [rawLiqData, crocEnv, chainId, currentPoolPriceTick]);
 
     return {
-        contextMatchesParams,
         baseTokenAddress,
         quoteTokenAddress,
         baseTokenDecimals, // Token contract decimals
