+++ src/contexts/TradeDataContext.tsx
@@ -9,6 +9,7 @@ import {
     useMemo,
     useState,
 } from 'react';
+import { useLocation } from 'react-router';
 import { getDefaultPairForChain } from '../ambient-utils/constants';
 import { isStablePair, translateTokenSymbol } from '../ambient-utils/dataLayer';
 import { TokenIF } from '../ambient-utils/types';
@@ -18,6 +19,7 @@ import { TokenContext } from './TokenContext';
 export interface TradeDataContextIF {
     tokenA: TokenIF;
     tokenB: TokenIF;
+    contextMatchesParams: boolean;
     baseToken: TokenIF;
     quoteToken: TokenIF;
     areDefaultTokensUpdatedForChain: boolean;
@@ -133,6 +135,33 @@ export const TradeDataContextProvider = (props: { children: ReactNode }) => {
               ? dfltTokenA
               : dfltTokenB,
     );
+
+    const pathname = useLocation().pathname;
+
+    // hook to sync token addresses in RTK to token addresses in RTK
+    const contextMatchesParams = useMemo(() => {
+        let matching = false;
+        const tokenAAddress = tokenA.address;
+        const tokenBAddress = tokenB.address;
+
+        if (pathname.includes('tokenA') && pathname.includes('tokenB')) {
+            const getAddrFromParams = (token: string) => {
+                const idx = pathname.indexOf(token);
+                const address = pathname.substring(idx + 7, idx + 49);
+                return address;
+            };
+            const addrTokenA = getAddrFromParams('tokenA');
+            const addrTokenB = getAddrFromParams('tokenB');
+            if (
+                addrTokenA.toLowerCase() === tokenAAddress.toLowerCase() &&
+                addrTokenB.toLowerCase() === tokenBAddress.toLowerCase()
+            ) {
+                matching = true;
+            }
+        }
+
+        return matching;
+    }, [pathname, tokenA, tokenB]);
 
     const [blackListedTimeParams, setBlackListedTimeParams] = useState<
         Map<string, Set<number>>
@@ -280,6 +309,7 @@ export const TradeDataContextProvider = (props: { children: ReactNode }) => {
     const tradeDataContext = {
         tokenA,
         tokenB,
+        contextMatchesParams,
         baseToken,
         quoteToken,
         isTokenABase,
